Rename 'aquantia' files to more suitable 'aqc1xx'.
