+++ src/include/ipxe/errfile.h
@@ -213,7 +213,7 @@ FILE_LICENCE ( GPL2_OR_LATER_OR_UBDL );
 #define ERRFILE_usbblk		     ( ERRFILE_DRIVER | 0x00ce0000 )
 #define ERRFILE_iphone		     ( ERRFILE_DRIVER | 0x00cf0000 )
 #define ERRFILE_slirp		     ( ERRFILE_DRIVER | 0x00d00000 )
-#define ERRFILE_aquantia	     ( ERRFILE_DRIVER | 0x00d00000 )
+#define ERRFILE_aqc1xx		     ( ERRFILE_DRIVER | 0x00d00000 )
 #define ERRFILE_atl_hw		     ( ERRFILE_DRIVER | 0x00d10000 )
 #define ERRFILE_atl2_hw		     ( ERRFILE_DRIVER | 0x00d20000 )
